SRTtoSYLT/requirements.txt
srt
# mutagen # for prototyp sylt processor
moviepy
pillow
matplotlib